restructure enums and add docstrings
