+++ python/ray/serve/_private/deployment_state.py
@@ -22,15 +22,12 @@ from ray.serve._private.autoscaling_state import AutoscalingStateManager
 from ray.serve._private.cluster_node_info_cache import ClusterNodeInfoCache
 from ray.serve._private.common import (
     DeploymentID,
-    DeploymentStatus,
     DeploymentStatusInfo,
     DeploymentStatusInternalTrigger,
-    DeploymentStatusTrigger,
     DeploymentTargetInfo,
     Duration,
     MultiplexedReplicaInfo,
     ReplicaID,
-    ReplicaState,
     RunningReplicaInfo,
 )
 from ray.serve._private.config import DeploymentConfig
@@ -70,6 +67,12 @@ from ray.serve.schema import (
     DeploymentDetails,
     ReplicaDetails,
     _deployment_info_to_schema,
+)
+from ray.serve.types import (
+    ALL_REPLICA_STATES,
+    DeploymentStatus,
+    DeploymentStatusTrigger,
+    ReplicaState,
 )
 from ray.util.placement_group import PlacementGroup
 
@@ -185,7 +188,6 @@ SLOW_STARTUP_WARNING_PERIOD_S = int(
     os.environ.get("SERVE_SLOW_STARTUP_WARNING_PERIOD_S", 30)
 )
 
-ALL_REPLICA_STATES = list(ReplicaState)
 _SCALING_LOG_ENABLED = os.environ.get("SERVE_ENABLE_SCALING_LOG", "0") != "0"
