+++ python/ray/serve/tests/unit/test_deployment_state.py
@@ -10,10 +10,7 @@ from ray.serve._private.autoscaling_state import AutoscalingStateManager
 from ray.serve._private.common import (
     DeploymentHandleSource,
     DeploymentID,
-    DeploymentStatus,
-    DeploymentStatusTrigger,
     ReplicaID,
-    ReplicaState,
     TargetCapacityDirection,
 )
 from ray.serve._private.config import DeploymentConfig, ReplicaConfig
@@ -49,6 +46,7 @@ from ray.serve._private.utils import (
     get_capacity_adjusted_num_replicas,
     get_random_string,
 )
+from ray.serve.types import DeploymentStatus, DeploymentStatusTrigger, ReplicaState
 from ray.util.placement_group import validate_placement_group
 
 # Global variable that is fetched during controller recovery that
